+++ Biblioteca/UI/Models/Recargo.cs
@@ -13,6 +13,6 @@ namespace UI.Models
 
         public int rec_estado { get; set; }
 
-        public decimal rec_monto { get; set; }
+        public int rec_monto { get; set; }
     }
 }
